refactor: modeltarget facade domain service
This commit moves part of the modeltarget facade over to using model
migration domain. This is being done in an effort to remove the
credential validation logic from this facade.

The credential validation logic needs to be removed so that model config
from mongo state can be removed from this service.

As part of this commit I have also swapped this facade over to the new
internal errors package.