feat(model): move GetModelUsers to the model service
Move GetModelUsers from the access service to the model service. Also
remove the check inside the state for the apiUser permissions. Instead
have a filterUsers argument that can be used to filter the users
returned if the user is not an admin.

Also remove the external users from the results of the function. In
retrospect this was an unneccery additional complexity. This is not
something we did in 3.x and it is not done in other places outside the
permissions service. The everyone@external user will now be displayed
since it indicates the level of access for all external users. If an
external user has explicity been granted access on the model it will
still appear as before. (only the implicit ones that got it through
everyone@external will disappear)