feat(machine): wire machine cloud instance set
This patch introduces the domain methods for setting a machine cloud
instance into the provisioner task.

An important note about the design is that we want controller workers to
go through the domain services when available, so in this case it meant
that the compute provisioner could directly make use of the domain
service. On the other hand, the container provisioner is running on the
machine agents where we don't have access to Dqlite and therefore has to
continue going through the API.
To make this design consistent, a new config was added to the
provisioner task (`GetMachineInstanceInfoSetter`) and the two different
manifolds (compute and container provisioners) inject the different
implementations.