feat: pool atomic context
The attomic context can be pooled to prevent the creation of objects
on the heap. There was originally a TODO about doing this, but until
the state base was finalized it was left to do until later. There
are a few services already using the atomic setup, so it feels like
we should do this before it's forgotten.

There should be no changes to the testing, this is just an
implementation detail and should be transparent to the tests.