refactor: remove provider from apiserver dependency
The apiserver shouldn't have any knowledge of a provider directly.
Instead this information should be accessed via the service. We
unfortunately won't be able to move *everything* to the service
layer (that's clearly way too much work) and should be done over
time. Instead, we should access the provider from one place, the
individual services within a service factory. When we then do
move the calls into the services, we don't have to replumb that change
through.

The added benefit is that we won't rely on the provider in the
apiserver and unable to remove it. This has happened in the past
and it prevents us from doing the right thing.

This leads to better correctness of the application. There is
one way!