feat: implement controller proxy
Inorder to wire up the apiserver/registration proxier we need
to use a provider for the controller model uuid. Getting from some
random service didn't feel right. So we now have a controller proxy
service (maybe it should be called the proxy service). This is
wired up and works in a clean way that we expect.