refactor: remove credential validator
Removes the creation and usage of a credential validator from the
credential service on the cloud facade.

We are removing this validation as it relies heavily on existing mongo
implementations and patterns that are not present in the DQlite world
anymore.

Also now that we have the provider tracker we want to use this work for
establishing providers and not that of environs.New. This is a
regression in Juju functionality and new credential validation will need
to be added back in at another stage.