refactor: remove credential validation on update
When updating a credential for a cloud on a controller we previously had
logic that would instanstiate a new environ with the credential that is
being proposed in the update. With this new environ we would go through
each model and perform a stock take of instances in the cloud and Juju's
state.

This design is very hard to carry forward in a Dqlite world as we don't
have a way to tell the provider tracker what credential to use and also
validation of a provider based on instances could get us unstuck as the
Juju model changes during the credential check.