feat: remove usages of model.Users
The model user collection is being removed in the move to 4.0. Usages of
it are being replaced with GetModelUsers on the model service and the
collection in mongo will soon be deleted.

As part of this, the model migration user doc is also removed from
state. This is used when checking if a user trying to access a migrated
model gets told it has been migrated or not. If the user had no
permissions they were not told. This check can no longer be performed.
TODOs have been put in the code and a card created on Jira to address
this.