feat(ports): get all opened ports for a machine
Add methods to the service + state layer, designed to replicate the
functionality of the MachinePortRanges state interface.

Examining the usages of this interface, you will see it is only used to
batch-get and cache the opened ports across all the units on a specific
machine.

So this interface can be replicated with a single method, one which gets
all then opened ports for a machine, and groups them by unit.

Implement just that