test: fix all the failing tests
The tests required some rework to ensure that we have the right
provider. The way that the registration suite is tested leaves
a lot to be desired. It's setting up a full API suite, when we could
just use a httptesting package and call the handler directly.

For now we hack around it.