fix: improve error messages for unsupported features
Error messages for supported features were intended to be
user-friendly, but they are still too abstract to easily understand.
They refer to the concept of "features", which is an internal Juju
construct and as such, should not be mentioned to users. This commit
attempts to improve the error messaging to make it more
straightforward.