refactor: moves keymanager state to controller
This change updates the keymanager domain to move the state from the
model database to that of controller database. This was done to get RI
onto both users and models and consider authorized keys an authz problem
of the controller.

As such the state layer now needs to take a model id.