feat: add method for ensuring unit_state row
Given that the actual state values in this table are non-nullable, it
makes sense to add a row if we need to up-front rather than making every
update an upsert.

Potentially, we could just call this at unit creation and not include it
in the call to setting the state at all.