fix(ports): ban conflicting port ranges for colocated units
It turns out we do not just disallow overlapping port ranges to be open
across a singular unit, but across any units colocated on the same net
node.

Fix UpdatePortRanges to do the stronger verification

NOTE: There is a slight change of behaviour compared to Mongo. In Mongo,
for some reason we disallow matching port ranges on co-located units,
despite the fact we allow them across endpoints of a unit.

We now allow matching (but not conflicting) port ranges across units.