fix: move permission check out of UpdatePermission
UpdatePermission took the apiUser as an arguemnt. In this it was unique
amoung the permission functions. The function also optionally adds a
user. This was incorrectly implemented to also add some none external
users. To replicate behaviour in 3.x external users should always be
added (if not present) and local users should need to already exist.

This change refactors this check to outside the call to update
permissions.

It also renames the method on state to UpdatePermission
instead of UpsertPermission.

It also changes the UpdatePermission function to always add non local
users if they are not present and removes the optional arguemnts that
used to control this. The creator of this user is everyone@external
instead of the apiUser as it was before this change. This is to keep it
consistant with the way external users are added on login.