chore(environ): check for empty availability zone before assigning
This commit uniformizes how we assing availability zones in hardware
characteristics. When empty string, then leave a nil pointer on hardware
characteristics.

This was found because on the new Dqlite domain we have a foreign key to
availability_zone when creating machine cloud instances, and an empty
string breaks this relation.