chore: remove charm post request
The charm upload request using POST was deprecated in 3.3, but we
wanted to wait until 4.0 to remove. The new method to upload charms
was via PUT (/model-:modeluuid/charms/:object).

I've still kept the GET request around for now, until I'm confident
that we don't need to service this request. Not only is the GET
request problematic from a Juju perspective, it's overly complicated.
The charm is opened in memory to read portions of the charm archive,
which can potentially lead to OOM errors or undexpected memory
pressures whith large charms. Fortunately the end point isn't well
documented, so many of the features don't seem well used.
We're moving towards if you want information about a charm, you
can download that charm and extract the information yourself.

This PR also removes the ability to fetch the charm icon. I've checked
with the dashboard team, and they no longer use this functionality and
source all images from charmhub. The correct approach to this would
be to add additional meta information whilst uploading (similar to
annotations), that could contain the icon if required.