feat: dual write charm during upload
For local charms to work, we need to dual write the charm data
during an upload (PUT request). The problem is that the charm
data is horribly convoluted for uploading a charm and could easily
be simpilfied. The problem is that we can't do it now, or at least
not yet, as we need to move the bulk of the read methods from state
on to the service methods. Once that's done and the authoritive
source of information is the service, we can sweep back around to
this and clean it up.