feat: move supported features to the application service
The supported features are actually charm based feature set, so
should live in the application service. Until we have full end to
end paths for models and charms, we can just hang this of from
the service for now.

We can start to drop the cloud and credential services in the
application domain, as we're slowly dropping the need to invoke
a provider directly.