chore(port): move verification into service layer
We fail to open/close port ranges on units if a conflicting port range
is open on any co-located units.

Co-located units are units which exist on the same net-node.

This fact is business logic. However, it was asserted in the state
layer.

Lift this logic out into the service.