feat: provider for credential
To work on a provider with a temporary credential, we need to
be able to fork the current provider to allow only working and
validate on a credential. The provider will then be thrown away
after the credential has been verified.

It's possible to just call this out of band, but we then loose
the ability to just have one location for the environs.

There are some limitations with doing this, like the onus is on the
use to correctly handle the provider and to relinquish the it once
it's done.