feat: remove charm call from application offers
The application offers requires just the charm description from
the charm when creating an application offer. We no longer need to
extract the whole charm, just the description. The prior code is
wild! If you have a large charm, you need to read the whole charm
just to get one field. This is changing, we just don't need to do
that.

This adds the state call to do this is. Also I'm going to remove
GetCharmByApplicationName, we'll just use the GetCharmIDByApplicationName.
This will then allow the usage of all the charm state commands, making
it a lot more useful.