chore: use machine cloud instance watcher in uniter facade
This patch introduces the use of the machine cloud instance watcher in
the uniter facade for lxd profiles. Since we are not implementing this
feature (lxd profiles) in dqlite at the moment, this watcher is not
supposed to do nothing and simply leave the scaffold for when the actual
watcher is implemented.