feat(port): add domain methods for app ports
We need two methods. One which simply returns opened ports by unit, by
endpoint for an application. This will be used for model migration

And another, which will be used by the caasfirewaller api, which groups
all opened port ranges on an application by endpoint only, and atomises
the port ranges.