fix: process peer relations last
It was found that when removing applications, hooks for some relations
rely on the application data populated by peer relations.

When peer relations are departed/broken first, these hooks fail,
blocking teardown.

It is reasonable to process existing relations peer-last to prevent this
situation. Hooks for relation creation and joining are unaffected.