chore: only one offer at a time
This forces all application offers to be added one at a time. The
client (via juju offer and the bundle handler) only allows adding
one offer at a time. The useful of the bulk API for this doesn't
seem well thought out, when concerning the scope of the facade call.
If the facade was scoped to the model, then adding multiple offers
for the scoped model, would make sense i.e. juju offer; juju offer.

Also, the juju offer command breaks convention by providing a new
way of targeting a model. Instead of using the conventional -m
command flag it uses the offer prefix to target a model.
We could fix the juju offer command to inherit from the model command
base and talk directly to the model facade instead of the controller
facade, but it breaks compatibility. The effort involved in changing
the API facade doesn't seem worth it.
Instead, we're going to limit the number of offers that can be added
at one time.

The main driver for this, is to ensure that our API is correct in
4.0, with respect of being able to access other model databases from
inside of a model facade. Model facades are supposed to work on the
model it's been created with. It is not expected to be able to
look into another database. The model might not even be on the same
controller, how do handle redirects etc. Another concern is ensuring
the correct permissions for each model in the request is correct for
every model being accessed. It's hard enough to get it right for
the same scoped model, never mind for every model being passed in.
There are other concerns around performance of accessing other models
from another model; potential denial of service (DOS) from another
model (one model can flood another model, using this to prevent a
model forward progress).

As you can see there are some concerns about an API like this, if
it's not totally thought out correctly.