feat: multi-model facade for offers
The offers facade method is wrong for a consistent view point. The
facade should be in the model scope, and access the controller model
when it needed to write there. Instead, we now have a bulk API that
allows the writing to different models from inside a model scope.
Nothing is taking advantage of the fact that it's a bulk API. I also
refute that a multi-model bulk API was a good idea ever. I could
live with a bulk API for a given model; add multiple offers for a
model at once.

We don't have the time and resources to unpick this atm. I might
undo this in the future, but not whilst undertaking charm changes.
I've now bitten the bullet and allowed the current model (controller
model), to view other models. The only restriction is that only 1
model can be accessed per facade call. This is to ensure that if we
ever want to restrict this call using rate-limiting, we can.

The code is really simple, register a multi-model context and grab
the service factory when we require it.

Lastly, I've removed all the model info calls, there is no need.
We can just rely on the checkAdmin call. It will either work or
wont. Destroying a model should remove the permissions and remove
it from the namespace_list. If that's not the case, we've got bigger
issues.