test: retry revision checking in secrets tests
Several secrets tests were failing when trying to check that old secret
revisions had been pruned. This process can take a couple of seconds,
but the tests were checking right away and hence failing. I've added a
check_num_secret_revisions function, which does the check in a retry
loop, so it is more resilient to timing issues.

Also, the check_* functions were using `exit 1` on failure, which
causes the whole process to exit (much like a panic in Go). Replace
this with `return 1`, so that the failure is recoverable.