feat: remove application charm from provisioner
The provisioner used application.Charm() for getting the LXDProfile.
We no longer require to access the whole charm just for the Profile.
The change just uses the application service to get the LXDProfile.
In order to ensure we have a unique LXDProfile per charm revision,
we also need the charm revision. Changing this in the service and
state code allows us to know what revision the LXDProfile is for.

Bespoke API requests on the service is providing lots of benefits.