fix: model migration
Now that we have real models and not just the controller model, it
was spotted that we weren't migrating any other model, other than
the controller model. This was a super easy fix, that I already
had part of from an old branch, which couldn't be applied at the
time because models didn't correctl exist.

The code is simple, just late bind to the model and state when
attempting an export.