fix(log): log block devices at trace level
There was a INFO log line coming from the diskmanager worker which
logged the full details of all block devices. This was creating a lot
of noise in the debug logs. This is far too much information for an
Infof call, it would be more appropriate at the Trace level.