refactor: remove model name checking in FindEntity
There was logic inside state.FindEntity that was checking if the model
tag used the model name instead of the UUID. It had to pull model
config to ensure the provided tag was equal to the current model's
name. Model tags have not used the model name since pre-Juju 1.21 (see
https://bugs.launchpad.net/juju-core/+bug/1257587), so I think it is
safe to remove this code.