fix: update to github.com/juju/retry v1.0.1
I updated juju/retry because when turning on or off jitter, it shouldn't
affect the average time spent delaying. The existing code dropped your
average wait time by ~50% because it picked a random number from 0 to
the desired delay. The new one does +/- 20%, but that means that on
average it is very close to the same time without jitter. Obviously
being 10% faster at 100ms is not the same as 10% slower at 2s. However,
this is still systemically closer to the same average than doing always
faster. And I don't think it is actually worth trying to average out the
rest.

I just didn't want jitter on/off to dramatically change the wait times.