feat: insert change_log_namespace with trigger
Ensuring that the change_log_namspace id and the trigger namespace
code int match is frought with danger. It's too easy to get it wrong
and then the whole changelog stream breaks.

The fix is to insert the change_log_namespace with the trigger at
the same time. There is only one source of truth.