feat: invert the primary key logic
After more discussions, we believe that the primary key should always
be checked. If the primary key is made of a composite key, then we
emit the value of the old row during the update, and it's up to the
watcher to check what the new value is. This is by design. It's so
that you can get both the prior value and the new value (subsequent
changes could have happened since you got the notification) during
the filter/mapper functions.

All the triggers now include the primary key for comparision checks.