feat: introduction of the objectstore services
The objectstore can not be embedded into the service factory because
the objectstore metadata services are inside the service factory. The
DAG of the dependencies prevent this, so to move forward, we can copy
what we've done with the provider tracker worker.

This introduces the concept for the workers. Adding it to the agent
manifolds and pushing it through the API is still required.