feat: reworkd objectstore dependencies
The objectstore dependencies have been changed to the objectstore
services (old terminology: object store service factory). This will
allow us to inject the objectstore into the service factory.

There is still a lot to fix, but according to the manifold tests,
we can clearly see that it's possible. So given that signal, I'm
going to press on and complete this.