feat(modelmigration): check provider and model machines instances match
This patch implements a missing step on the checks when migrating a
model, left as a TODO by @tlm.

Basically the check makes sure that the list of instance IDs retrieved
from the provider and the ones retrieved from the DB (model machines)
are exactly the same, any discrepancy results in a precheck error.