feat: transaction logging
The following allows the logging of all queries through the system.
Originally, I intended to shim out the sqlair.TX and the sql.Tx
packages, but unfortunately we can't get hold of the parsed query
in the sqlair.TX. The package abstracts away that part. No database
or statement indirection.

It then occured to me, that we already have the query in the OTEL
tracing. This was done by me when plumbing the tracing in the
dqlite library. The library is always tracing, we just push through
a no-op tracer if it's not required (stays on the stack!). Thus,
by aligning to the tracing interface, we can log the query to the
logger. To prevent excessive heap allocations, we pool the log tracer
and reuse as much as possible. The semaphore prevents this from growing
too large and reuse as much as possible.

A txnID was added to group transactions together, although we don't
get the number of retries. We can always thread that through at
a later date.

Although the txn runner is getting more complicated, it is doing
exactly what we want and containing the complexity of interfacing
with dqlite really well.