feat: add model config provider
This commit adds a new interface type to the environ package to describe
providers that want to have opinions about a models configuration. I am
introducing this change to place together all of the disparate funcs we
currently have that mix concerns.

This will make it easier to inject into the model defaults domain.