feat: move agent object store
The controller also has a object store, that also needs to be
migrated across to the new object store services.

The service factory (soon to be renamed) is becoming lighter and
is only offering the services that are serviable by the API
server.

The domain services are in fact real services, not infrastructure.