chore: final wireup of the manifolds
The following wires up all the service factory dependencies correctly.
It's now possible to have the object store in the service factory.

There is a question around if the service factory should have a
deprendency on the upgrade database. Originally it didn't because
it's lazy evaluated. Now it's been introduced by the objectstore.

I wonder if that's going to cause us issues.