fix: model migration failure due to model config not found
In the process of removing model config from state, there is
functionality which needs to be skipped for now until it can be
re-implemented in a domain. In this case skip validation model
constraints for now. The model domain will be completed after model
config.

Skip 2 tests which fail due to above change. Remove once they have been
handled in the model epic.