refactor: authorized keys message on add model
In previous PR's we have removed the ability to add juju ssh keys when
adding a model due to security impacts. This left an ugly error message
users would see when adding models.

Now when adding a model we simply inform the user of what they need to
do from there onwards.