feat: provider defaults in model config
This commit adds support get getting provider based model config
defaults in the DQlite domain services. Previously this was only
happening for config attributes that the provider is extending model
config with and not defaults the provider wanted to apply for controller
wide attributes. The most common case of a provider wanting to do this
is for storage.

As well in this PR we are know longer letting the mongo based model
config creation process take place. Instead we are using the DQlite
domain services to perform the full generation and copying this result
down into Mongo while we dual write.

Because of this I have been able to remove the controller/modelmanager
package that is no longer needed.