feat(lxdprofiles): add lxd profiles watcher
Before this patch, the lxd watcher on the uniter facades used the
instance data watcher underneath. This was suboptimal because a lot of
unrelated changes were triggered.

With the migration to dqlite, we now add a new lxd profiles watcher
which replaces the legacy instance data watcher on the uniter.

Bonus: The machine cloud instance watcher is now transformed into a
notify watcher and taking a machine UUID as argument, to match what's
needed on the call site.

Bonus 2: Small fix to use the FilterEvents method on the machine reboots
watcher.