feat: allow a precheck per transaction
The precheck concept is to run a function for every retry to ensure
that something is as close to being true, without being run in a
transaction. This might be something like; ensuring a lease exists
for a given unit, or running another query from another database,
without running it within the same transaction. Nested transactions
are undefined behaviour and we should do our best not to allow that.