refactor: remove std txn from domain
We shouldn't be using StdTxn from inside the domain package. It's
fine with in tests, but all queries should be using sqlair.

Reducing the interface of the TxnRunner then allows to add a
TxnWithPrecheck to prevent bloating out the TxnRunner. We want to
clean up and ensure clarity whilst we go.