feat: signal when the lease wait has started
In order to ensure that the lease is correctly being held we need
to wait, then check. That ensures that there is no race if we do
it the other way around. By telling us when it's started, we can
the guard the check on that before running.

Sequencing then becomes easy.