feat: check lease whilst running transaction
This introduces the ability to check that we have the approriate
lease when running a transaction. The context.Context that is passed
to the closure will be cancelled if the leader changes from the
original holder.

The WaitUntilExpired signature was changed, so that we can guarantee
that we're starting the wait. This ensures that there are no races
between checking the token and then calling WaitUntilExpired.

The call to token is done first to ensure that we don't perform any
additional operations that aren't required.

The go routines are sequenced to ensure that we first block and wait,
then check that the lease token. If that matches what we expect, then
and only then do we run the function.

If the WaitUntilExpired unblocks, we'll cancel the context and everything
must be stopped at the first possible moment. This includes rolling
back any in-play transactions.