fix: actually only register 1 metrics endpoint
The issue really was that we were registering the same (Name,Subsystem) multiple times.
The only way around this is to push up the registration of the perf worker higher, and
have it create a CounterVec. We can re-use ForModel. I don't know that it is ideal,
as it means the agent/engine knows about additional metrics that are specific to 1 worker.
But we're just hacking it together here.