fix: destroy-model defaults to waiting 2s
When creating a lot of models and then destroying them, it turns out that
we just had a direct 'always wait 2s'. Occasionally when destroying 20 models
the model would be cleaned up by the time we looked, but most of the time it
takes a short time to actually cleanup.
This shouldn't impact most people, as waiting for real machines to tear down,
it will quickly ramp up to waiting 2s again.
However, when doing testing, this allows me to teardown 20 models in 7s instead of
taking 40s.