feat: remove semaphore use driver pool
This change is the first in a set of changes to improve the performance
of transactions in dqlite. Prior to this change, it was expected
that dqlite could just scale because it felt faster. Except when
running tests and comparing the performance against 3.6, the 3.6
version was miles ahead.

Lots of testing was done[1] to work out where our bottlenecks occured.

One of the first outcomes from the testing was that it's better to
offload the semaphore to the database/sql pool. Originally, it was
expected that the semaphore would prevent additional work being
performed. In then end, that was a fallacy.

Documentation[2] from go even states:

    Keep in mind that setting a limit makes database usage similar
    to acquiring a lock or semaphore, with the result that your
    application can deadlock waiting for a new database connection.

This is a simple change and a nice win.

The number used here is currently set to 2, but once we implement
a dqlite database ring, that will be changed to a higher number as
it unlocks better performance once we're sharding.

 1. https://github.com/juju/juju/pull/18007
 2. https://go.dev/doc/database/manage-connections