docs: improve comment on db connections
We have to be very careful what we set the database connections to,
if it's too large we start to create lots of connections. This is
fine as they eventually get reused, but for small workloads that's
probably not ideal. For large setups, the more the better.

For now I've set it to 2, because testing has shown it's marginally
better that 1, because of queue dynamics. The next pool item is
ready to go, whilst the initial one is finishing. Thrashing over
the pool is a concern though.