fix: ignore ModelNotFound in LastestMigration if Done
LatestMigration is called by SetPhase. There is a bug where if the phase
is done then the "if" block will be triggered and the deleted model not
found, causing LatestMigration to return an error.

This was stopping post-migration steps such as setting up offer
redirects from happening.

Change this to be idempotent by ignoring the NotFound error and making
the code more resiliant.