chore: remove all state tests
To keep the unit tests green, we spend time on updating the old
tests, that are just going to be thrown away. This is one of
the first steps in reducing the burden on migrating to dqlite.

It is expected that we will refer to the 3.6 branch for defining
or referring to epic creation. Juju developers will need to refer
to 3.6 branch to ensure compatibility and consistency.