feat(unit): add GetUnitUUID method to service
This will be required to wire up other domains. The pattern we're
following is to use natural keys for resources over the api, and
resolving into uuid at the apiserver.

To do this, we will need a method to convert a unit name into it's UUID.
We already have a state method for this. Expose it on the service.

Also, refactor the domain slightly to make use of static unit and
application id types in method signatures.