fix: checking for hooks in local charm
When checking if a local charm has hooks, we were assuming that the
zip file would always have a separate entry for the directory 'hooks/'
and for each of its subfiles. However, this is not the case for charms
generated by Charmcraft. A better check is that there exists at least
one file in the 'hooks' directory that itself is not a directory.