feat(application): add GetUnitName method
Add a method to the application domain which translates the uuid of a
unit into it's name.

This will be useful at the edges of our APIs. The pattern we are
intending to follow is to use natural keys in the API contract, but to
convert to uuids internally.

We already have a method to convert the natural key of a unit into the
uuid. But we also need one to do the reverse. This will be used to
return natural keys in API responses from uuids.