docs: link to juju specific conventional commits info
Currently the initial link only goes to the conventional commits
website, which leaves the type field of conventional commits
underspeficied. The types used in this project are linked to later in
this document.

Change the initial link go to this later section, which can explicitly
link to both the general conventional commits information and the juju
project specific guidelines.