test: skip destroy for smoke tests
Smoke tests for localhost should be skipped whilst we start dual
writing for units for cleanup. There is a logical race in the
watchers that it preventing the cleanup of the machines.

This is only for the github action and is default to off, we need
this to keep testing bootstrap otherwise we can introduce bugs.

This should be removed once clean up is done.