feat(stub): implement a stub service
We have taken the decision to implement a special service called
"Stub", which will collect together temporary methods required to wire
up domains which depends on domains which have not been entirely
implemented or wired up.

Implement the first method on this service, AssignUnitsToMachines. As of
writing, the application and machine services do nbot work together to
assign units to machines. Instead, new stub uuids are always generated
and added to new units and machines.

This means units are never assigned to machined in DQLite. But this is
required to wire up the ports domain.

So wire up this new method, which moves units to machine, whenever new
units are added.

Since this is just a temporary solution, we do not need to worry about
stale net nodes.