fix: prevent uniter bouncing
The controller uniter is bouncing because it's trying to get hold
of the az, but it can't be found, which isn't the error that the
uniter is expecting. It is expecting a not provisioned error.

This code is not correct, because we've not got the correct errors
in place for the machine domain.

For now this allows us to continue with out progress.