feat(firewall): wire up firewaller to use ports domain
Wire up the firewaller api on the controller facade to be backed by
the DQLite ports domain.

This included substituting the DQLite watcher for the Mongo watcher.

This required a quick patch to the DQLite watcher to emit machine names
instead of machine uuids, since this data is consumed outside the
apiserver, so should be a natural key.

We also needed to wire up the OpenedPortRangesForMachine facade
endpoint. This required a bit of hacking (i.e. hardcoding SubnetCIDRs to
nil) to wire up. This was done since the application endpoints domain is
not yet wired up, so is still backed by Mongo on this facade.

This makes it very difficult to navigate from a DQLite unit to that
unit's endpoint bindings in Mongo, to the point it's not worth doing.
Instead, we will restore this once endpoint bindings are backed by
DQLite.

Restoring this has been added to the risk register.