test: fix local charm integration tests
The local charm integration tests relied on the ntp charm which was not
up to date with the latest charm requrements and caused an error on
deploy. It was held in launchpad and not under our control.

Switch to using the ubuntu-plus charm instead from test testcharms
directory.

Note that these tests deploy from directory so will not work in 4.0.
They will need to be updated or maybe even just deleted when the patch
is merged up.