chore(instancedata): full cleanup of instance data on legacy state
This patch removes the last pieces of instance data from the legacy
state (the collection and some bits of the
methods `SetInstanceInfo` and `SetProvisioned`).

The only non-removal code is a fix in the
provisioner_integration_test.go, where one test was still calling
SetProvisioned from the legacy state. This was easily replaced by the
service method.