feat: use read only model info rather than config
Replacing calles to model.Config and st.ModelConfig with the ModelInfoService
where the agent version and cloud type values are needed. Otherwise the
ModelConfigService would be used.

As I'm unable to remove the Config and ModelConfig methods from all
state interfaces during this commit, cause a panic if the stub methods
are used in tests.