fix(migration): move description model validation after domain
Currently, after filling the `description.Model` when migrating the
legacy db contents we validate it. This happens before we start filling
it with the contents from dqlite.

We face an issue here, that on some cases the validation will fail
because some of the missing data will arrive after we finish the dqlite
stage of the migration. This is the case for machines/instance data
migration.

We therefore move the whole Validation() to internal/migration, *after*
dqlite.