feat(dqlite): move cloud image metadata from model databases to controller db
- Before this commit, image metadata handling was split between the model and the controller.
- After this commit, all image metadata handling is centralized in the controller.

This change is a rollback from previous decision where we decided that cloud image metadata can be
handled separately in each model, as the documentation of the plugin `juju metadata` suggested.

This would have been a possible breaking change from 3.6 and we decided to not change this behaviour
for now. So, after this commit, it should behave the same way it behaves in 3.6, regarding
bootstrap, provisioning and migration.

Signed-off-by: Guillaume Fouillet <guillaume.fouillet@canonical.com>